Start adding integration tests on app (#49)
* setup jest

* split app creation and app launch

* Add tests for file and history endpoints

* fix package-lock.json

* update package-lock

* add sinon as an explicit dependency
