+++ package.json
@@ -53,6 +53,7 @@
     "autoprefixer": "^10.3.1",
     "jest": "^27.0.6",
     "node-stream-zip": "^1.14.0",
+    "sinon": "^11.1.2",
     "supertest": "^6.1.6",
     "ts-jest": "^27.0.5",
     "ts-mock-imports": "^1.3.7",
